Implement Packer-native concurrent builds
HCL2 appears to lack the expressiveness to generate top level blocks on
its own, such as "source" and "build".  While it does have some loop
support, it appears limited to generating collections.  Had to create a
jinja2 template for the HCL2 template in order to get all the builds
expressed simultaneously and allow Packer to utilize its parallelism.

There is a new "inv template" task that can be run manually, and it's
also hooked as a pre-task to all routeros build tasks.

Also decided to change the initialization of the context to prepopulate
it with the latest RouterOS versions from Mikrotik.  This data is used
throughout the build program, and it was being fetched in multiple
redundant places.  Now it's centralized and always available.

Had to change the box filenames to include a provider in the middle.
The new syntax is: <branch>_<provider>_<version>.box.  Since we're
supporting multiple build providers now, the filename distinction needs
to be made.

Dropped the generation of the associated *.md files along with the
boxes.  Not sure what value it provided, since the boxes already carry
their own unique version information.

Finally, removed the ask_for_confirmation when invoking a build.  If
someone didn't want to build, they wouldn't have invoked the command.

Future note: if the build host is equipped with both libvirt and
virtualbox, it may be possible to run builds in both systems
concurrently.  This would produce all possible boxes in the shortest
amount of time.  The current logic of selecting libvirt in favor of
virtualbox may need to be changed to "select all available providers".