Migrate the plugin build process out of Vagrant
This is many times faster than with a Vagrant build VM.  The dependency on
Ruby 2.7 is also eliminated with this change.  Removed manual build
instructions and replaced with a reference to more granular build tasks.
The new build system uses a localized (not global) bundle installation.
This will keep the global build environment clean.  Global build requirements
have been updated in the README.md.  Debian/Ubuntu build environment
instructions should be possible, but are not provided at this time.